feat(evm-core): gate optimism behind cargo feature (#14593)
