+++ crates/evm/core/Cargo.toml
@@ -19,7 +19,7 @@ foundry-common.workspace = true
 foundry-config.workspace = true
 foundry-evm-abi.workspace = true
 foundry-evm-hardforks.workspace = true
-foundry-evm-networks = { workspace = true, features = ["optimism"] }
+foundry-evm-networks.workspace = true
 
 alloy-chains.workspace = true
 alloy-dyn-abi = { workspace = true, features = ["arbitrary", "eip712"] }
@@ -36,7 +36,7 @@ alloy-primitives = { workspace = true, features = [
 alloy-provider.workspace = true
 alloy-network.workspace = true
 alloy-consensus.workspace = true
-alloy-op-evm.workspace = true
+alloy-op-evm = { workspace = true, optional = true }
 alloy-rpc-types = { workspace = true, features = ["anvil"] }
 alloy-sol-types.workspace = true
 alloy-rlp.workspace = true
@@ -54,9 +54,9 @@ revm = { workspace = true, features = [
     "blst",
 ] }
 revm-inspectors.workspace = true
-op-alloy-consensus = { workspace = true, features = ["k256"] }
-op-alloy-network.workspace = true
-op-revm.workspace = true
+op-alloy-consensus = { workspace = true, features = ["k256"], optional = true }
+op-alloy-network = { workspace = true, optional = true }
+op-revm = { workspace = true, optional = true }
 tempo-revm.workspace = true
 tempo-alloy.workspace = true
 tempo-contracts.workspace = true
@@ -81,3 +81,14 @@ op-alloy-consensus.workspace = true
 op-alloy-rpc-types.workspace = true
 anvil.workspace = true
 foundry-test-utils.workspace = true
+
+[features]
+default = ["optimism"]
+optimism = [
+    "dep:op-alloy-consensus",
+    "dep:op-alloy-network",
+    "dep:alloy-op-evm",
+    "dep:op-revm",
+    "foundry-evm-hardforks/optimism",
+    "foundry-evm-networks/optimism",
+]
