+++ crates/evm/core/src/env.rs
@@ -4,13 +4,9 @@ use alloy_consensus::Typed2718;
 pub use alloy_evm::EvmEnv;
 use alloy_evm::FromRecoveredTx;
 use alloy_network::{AnyRpcTransaction, AnyTxEnvelope, TransactionResponse};
-use alloy_op_evm::OpTx;
 use alloy_primitives::{Address, B256, Bytes, U256};
-use op_alloy_consensus::{DEPOSIT_TX_TYPE_ID, TxDeposit};
-use op_revm::{
-    OpTransaction,
-    transaction::{OpTxTr, deposit::DEPOSIT_TRANSACTION_TYPE},
-};
+#[cfg(feature = "optimism")]
+use op_revm::transaction::deposit::DEPOSIT_TRANSACTION_TYPE;
 use revm::{
     Context, Database, Journal,
     context::{Block, BlockEnv, Cfg, CfgEnv, Transaction, TxEnv},
@@ -236,9 +232,16 @@ pub trait FoundryTransaction: Transaction {
     /// Sets whether the transaction is a system transaction
     fn set_system_transaction(&mut self, _is_system_transaction: bool) {}
 
-    /// Returns `true` if transaction is of type [`DEPOSIT_TRANSACTION_TYPE`].
+    /// Returns `true` if transaction is an Optimism deposit transaction.
     fn is_deposit(&self) -> bool {
-        self.tx_type() == DEPOSIT_TRANSACTION_TYPE
+        #[cfg(feature = "optimism")]
+        {
+            self.tx_type() == DEPOSIT_TRANSACTION_TYPE
+        }
+        #[cfg(not(feature = "optimism"))]
+        {
+            false
+        }
     }
 
     // Tempo methods
@@ -317,188 +320,6 @@ impl FoundryTransaction for TxEnv {
 
     fn set_max_fee_per_blob_gas(&mut self, max_fee_per_blob_gas: u128) {
         self.max_fee_per_blob_gas = max_fee_per_blob_gas;
-    }
-}
-
-impl<TX: FoundryTransaction> FoundryTransaction for OpTransaction<TX> {
-    fn set_tx_type(&mut self, tx_type: u8) {
-        self.base.set_tx_type(tx_type);
-    }
-
-    fn set_caller(&mut self, caller: Address) {
-        self.base.set_caller(caller);
-    }
-
-    fn set_gas_limit(&mut self, gas_limit: u64) {
-        self.base.set_gas_limit(gas_limit);
-    }
-
-    fn set_gas_price(&mut self, gas_price: u128) {
-        self.base.set_gas_price(gas_price);
-    }
-
-    fn set_kind(&mut self, kind: TxKind) {
-        self.base.set_kind(kind);
-    }
-
-    fn set_value(&mut self, value: U256) {
-        self.base.set_value(value);
-    }
-
-    fn set_data(&mut self, data: Bytes) {
-        self.base.set_data(data);
-    }
-
-    fn set_nonce(&mut self, nonce: u64) {
-        self.base.set_nonce(nonce);
-    }
-
-    fn set_chain_id(&mut self, chain_id: Option<u64>) {
-        self.base.set_chain_id(chain_id);
-    }
-
-    fn set_access_list(&mut self, access_list: AccessList) {
-        self.base.set_access_list(access_list);
-    }
-
-    fn authorization_list_mut(
-        &mut self,
-    ) -> &mut Vec<Either<SignedAuthorization, RecoveredAuthorization>> {
-        self.base.authorization_list_mut()
-    }
-
-    fn set_gas_priority_fee(&mut self, gas_priority_fee: Option<u128>) {
-        self.base.set_gas_priority_fee(gas_priority_fee);
-    }
-
-    fn set_blob_hashes(&mut self, _blob_hashes: Vec<B256>) {}
-
-    fn set_max_fee_per_blob_gas(&mut self, _max_fee_per_blob_gas: u128) {}
-
-    fn enveloped_tx(&self) -> Option<&Bytes> {
-        OpTxTr::enveloped_tx(self)
-    }
-
-    fn set_enveloped_tx(&mut self, bytes: Bytes) {
-        self.enveloped_tx = Some(bytes);
-    }
-
-    fn source_hash(&self) -> Option<B256> {
-        OpTxTr::source_hash(self)
-    }
-
-    fn set_source_hash(&mut self, source_hash: B256) {
-        if self.tx_type() == DEPOSIT_TRANSACTION_TYPE {
-            self.deposit.source_hash = source_hash;
-        }
-    }
-
-    fn mint(&self) -> Option<u128> {
-        OpTxTr::mint(self)
-    }
-
-    fn set_mint(&mut self, mint: u128) {
-        if self.tx_type() == DEPOSIT_TRANSACTION_TYPE {
-            self.deposit.mint = Some(mint);
-        }
-    }
-
-    fn is_system_transaction(&self) -> bool {
-        OpTxTr::is_system_transaction(self)
-    }
-
-    fn set_system_transaction(&mut self, is_system_transaction: bool) {
-        if self.tx_type() == DEPOSIT_TRANSACTION_TYPE {
-            self.deposit.is_system_transaction = is_system_transaction;
-        }
-    }
-}
-
-impl FoundryTransaction for OpTx {
-    fn set_tx_type(&mut self, tx_type: u8) {
-        self.0.set_tx_type(tx_type);
-    }
-
-    fn set_caller(&mut self, caller: Address) {
-        self.0.set_caller(caller);
-    }
-
-    fn set_gas_limit(&mut self, gas_limit: u64) {
-        self.0.set_gas_limit(gas_limit);
-    }
-
-    fn set_gas_price(&mut self, gas_price: u128) {
-        self.0.set_gas_price(gas_price);
-    }
-
-    fn set_kind(&mut self, kind: TxKind) {
-        self.0.set_kind(kind);
-    }
-
-    fn set_value(&mut self, value: U256) {
-        self.0.set_value(value);
-    }
-
-    fn set_data(&mut self, data: Bytes) {
-        self.0.set_data(data);
-    }
-
-    fn set_nonce(&mut self, nonce: u64) {
-        self.0.set_nonce(nonce);
-    }
-
-    fn set_chain_id(&mut self, chain_id: Option<u64>) {
-        self.0.set_chain_id(chain_id);
-    }
-
-    fn set_access_list(&mut self, access_list: AccessList) {
-        self.0.set_access_list(access_list);
-    }
-
-    fn authorization_list_mut(
-        &mut self,
-    ) -> &mut Vec<Either<SignedAuthorization, RecoveredAuthorization>> {
-        self.0.authorization_list_mut()
-    }
-
-    fn set_gas_priority_fee(&mut self, gas_priority_fee: Option<u128>) {
-        self.0.set_gas_priority_fee(gas_priority_fee);
-    }
-
-    fn set_blob_hashes(&mut self, _blob_hashes: Vec<B256>) {}
-
-    fn set_max_fee_per_blob_gas(&mut self, _max_fee_per_blob_gas: u128) {}
-
-    fn enveloped_tx(&self) -> Option<&Bytes> {
-        FoundryTransaction::enveloped_tx(&self.0)
-    }
-
-    fn set_enveloped_tx(&mut self, bytes: Bytes) {
-        self.0.set_enveloped_tx(bytes);
-    }
-
-    fn source_hash(&self) -> Option<B256> {
-        FoundryTransaction::source_hash(&self.0)
-    }
-
-    fn set_source_hash(&mut self, source_hash: B256) {
-        self.0.set_source_hash(source_hash);
-    }
-
-    fn mint(&self) -> Option<u128> {
-        FoundryTransaction::mint(&self.0)
-    }
-
-    fn set_mint(&mut self, mint: u128) {
-        self.0.set_mint(mint);
-    }
-
-    fn is_system_transaction(&self) -> bool {
-        FoundryTransaction::is_system_transaction(&self.0)
-    }
-
-    fn set_system_transaction(&mut self, is_system_transaction: bool) {
-        self.0.set_system_transaction(is_system_transaction);
     }
 }
 
@@ -687,32 +508,6 @@ impl FromAnyRpcTransaction for TxEnv {
     }
 }
 
-impl FromAnyRpcTransaction for OpTx {
-    fn from_any_rpc_transaction(tx: &AnyRpcTransaction) -> eyre::Result<Self> {
-        if let Some(envelope) = tx.as_envelope() {
-            return Ok(Self(OpTransaction::<TxEnv> {
-                base: TxEnv::from_recovered_tx(envelope, tx.from()),
-                enveloped_tx: None,
-                deposit: Default::default(),
-            }));
-        }
-
-        // Handle OP deposit transactions from `Unknown` envelope variant.
-        if let AnyTxEnvelope::Unknown(unknown) = &*tx.inner.inner
-            && unknown.ty() == DEPOSIT_TX_TYPE_ID
-        {
-            let mut fields = unknown.inner.fields.clone();
-            fields.insert("from".to_string(), serde_json::to_value(tx.from())?);
-            let deposit_tx: TxDeposit = fields
-                .deserialize_into()
-                .map_err(|e| eyre::eyre!("failed to deserialize deposit tx: {e}"))?;
-            return Ok(Self::from_recovered_tx(&deposit_tx, deposit_tx.from));
-        }
-
-        eyre::bail!("cannot convert unknown transaction type to OpTransaction")
-    }
-}
-
 impl FromAnyRpcTransaction for TempoTxEnv {
     fn from_any_rpc_transaction(tx: &AnyRpcTransaction) -> eyre::Result<Self> {
         use alloy_consensus::Transaction as _;
@@ -747,6 +542,222 @@ impl FromAnyRpcTransaction for TempoTxEnv {
     }
 }
 
+#[cfg(feature = "optimism")]
+mod optimism {
+    use super::*;
+    use alloy_op_evm::OpTx;
+    use op_alloy_consensus::{DEPOSIT_TX_TYPE_ID, TxDeposit};
+    use op_revm::{OpTransaction, transaction::OpTxTr};
+
+    impl<TX: FoundryTransaction> FoundryTransaction for OpTransaction<TX> {
+        fn set_tx_type(&mut self, tx_type: u8) {
+            self.base.set_tx_type(tx_type);
+        }
+
+        fn set_caller(&mut self, caller: Address) {
+            self.base.set_caller(caller);
+        }
+
+        fn set_gas_limit(&mut self, gas_limit: u64) {
+            self.base.set_gas_limit(gas_limit);
+        }
+
+        fn set_gas_price(&mut self, gas_price: u128) {
+            self.base.set_gas_price(gas_price);
+        }
+
+        fn set_kind(&mut self, kind: TxKind) {
+            self.base.set_kind(kind);
+        }
+
+        fn set_value(&mut self, value: U256) {
+            self.base.set_value(value);
+        }
+
+        fn set_data(&mut self, data: Bytes) {
+            self.base.set_data(data);
+        }
+
+        fn set_nonce(&mut self, nonce: u64) {
+            self.base.set_nonce(nonce);
+        }
+
+        fn set_chain_id(&mut self, chain_id: Option<u64>) {
+            self.base.set_chain_id(chain_id);
+        }
+
+        fn set_access_list(&mut self, access_list: AccessList) {
+            self.base.set_access_list(access_list);
+        }
+
+        fn authorization_list_mut(
+            &mut self,
+        ) -> &mut Vec<Either<SignedAuthorization, RecoveredAuthorization>> {
+            self.base.authorization_list_mut()
+        }
+
+        fn set_gas_priority_fee(&mut self, gas_priority_fee: Option<u128>) {
+            self.base.set_gas_priority_fee(gas_priority_fee);
+        }
+
+        fn set_blob_hashes(&mut self, _blob_hashes: Vec<B256>) {}
+
+        fn set_max_fee_per_blob_gas(&mut self, _max_fee_per_blob_gas: u128) {}
+
+        fn enveloped_tx(&self) -> Option<&Bytes> {
+            OpTxTr::enveloped_tx(self)
+        }
+
+        fn set_enveloped_tx(&mut self, bytes: Bytes) {
+            self.enveloped_tx = Some(bytes);
+        }
+
+        fn source_hash(&self) -> Option<B256> {
+            OpTxTr::source_hash(self)
+        }
+
+        fn set_source_hash(&mut self, source_hash: B256) {
+            if self.tx_type() == DEPOSIT_TRANSACTION_TYPE {
+                self.deposit.source_hash = source_hash;
+            }
+        }
+
+        fn mint(&self) -> Option<u128> {
+            OpTxTr::mint(self)
+        }
+
+        fn set_mint(&mut self, mint: u128) {
+            if self.tx_type() == DEPOSIT_TRANSACTION_TYPE {
+                self.deposit.mint = Some(mint);
+            }
+        }
+
+        fn is_system_transaction(&self) -> bool {
+            OpTxTr::is_system_transaction(self)
+        }
+
+        fn set_system_transaction(&mut self, is_system_transaction: bool) {
+            if self.tx_type() == DEPOSIT_TRANSACTION_TYPE {
+                self.deposit.is_system_transaction = is_system_transaction;
+            }
+        }
+    }
+
+    impl FoundryTransaction for OpTx {
+        fn set_tx_type(&mut self, tx_type: u8) {
+            self.0.set_tx_type(tx_type);
+        }
+
+        fn set_caller(&mut self, caller: Address) {
+            self.0.set_caller(caller);
+        }
+
+        fn set_gas_limit(&mut self, gas_limit: u64) {
+            self.0.set_gas_limit(gas_limit);
+        }
+
+        fn set_gas_price(&mut self, gas_price: u128) {
+            self.0.set_gas_price(gas_price);
+        }
+
+        fn set_kind(&mut self, kind: TxKind) {
+            self.0.set_kind(kind);
+        }
+
+        fn set_value(&mut self, value: U256) {
+            self.0.set_value(value);
+        }
+
+        fn set_data(&mut self, data: Bytes) {
+            self.0.set_data(data);
+        }
+
+        fn set_nonce(&mut self, nonce: u64) {
+            self.0.set_nonce(nonce);
+        }
+
+        fn set_chain_id(&mut self, chain_id: Option<u64>) {
+            self.0.set_chain_id(chain_id);
+        }
+
+        fn set_access_list(&mut self, access_list: AccessList) {
+            self.0.set_access_list(access_list);
+        }
+
+        fn authorization_list_mut(
+            &mut self,
+        ) -> &mut Vec<Either<SignedAuthorization, RecoveredAuthorization>> {
+            self.0.authorization_list_mut()
+        }
+
+        fn set_gas_priority_fee(&mut self, gas_priority_fee: Option<u128>) {
+            self.0.set_gas_priority_fee(gas_priority_fee);
+        }
+
+        fn set_blob_hashes(&mut self, _blob_hashes: Vec<B256>) {}
+
+        fn set_max_fee_per_blob_gas(&mut self, _max_fee_per_blob_gas: u128) {}
+
+        fn enveloped_tx(&self) -> Option<&Bytes> {
+            FoundryTransaction::enveloped_tx(&self.0)
+        }
+
+        fn set_enveloped_tx(&mut self, bytes: Bytes) {
+            self.0.set_enveloped_tx(bytes);
+        }
+
+        fn source_hash(&self) -> Option<B256> {
+            FoundryTransaction::source_hash(&self.0)
+        }
+
+        fn set_source_hash(&mut self, source_hash: B256) {
+            self.0.set_source_hash(source_hash);
+        }
+
+        fn mint(&self) -> Option<u128> {
+            FoundryTransaction::mint(&self.0)
+        }
+
+        fn set_mint(&mut self, mint: u128) {
+            self.0.set_mint(mint);
+        }
+
+        fn is_system_transaction(&self) -> bool {
+            FoundryTransaction::is_system_transaction(&self.0)
+        }
+
+        fn set_system_transaction(&mut self, is_system_transaction: bool) {
+            self.0.set_system_transaction(is_system_transaction);
+        }
+    }
+
+    impl FromAnyRpcTransaction for OpTx {
+        fn from_any_rpc_transaction(tx: &AnyRpcTransaction) -> eyre::Result<Self> {
+            if let Some(envelope) = tx.as_envelope() {
+                return Ok(Self(OpTransaction::<TxEnv> {
+                    base: TxEnv::from_recovered_tx(envelope, tx.from()),
+                    enveloped_tx: None,
+                    deposit: Default::default(),
+                }));
+            }
+
+            // Handle OP deposit transactions from `Unknown` envelope variant.
+            if let AnyTxEnvelope::Unknown(unknown) = &*tx.inner.inner
+                && unknown.ty() == DEPOSIT_TX_TYPE_ID
+            {
+                let mut fields = unknown.inner.fields.clone();
+                fields.insert("from".to_string(), serde_json::to_value(tx.from())?);
+                let deposit_tx: TxDeposit = fields
+                    .deserialize_into()
+                    .map_err(|e| eyre::eyre!("failed to deserialize deposit tx: {e}"))?;
+                return Ok(Self::from_recovered_tx(&deposit_tx, deposit_tx.from));
+            }
+
+            eyre::bail!("cannot convert unknown transaction type to OpTransaction")
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use std::num::NonZeroU64;
@@ -755,14 +766,10 @@ mod tests {
     use alloy_consensus::{Sealed, Signed, TxEip1559, transaction::Recovered};
     use alloy_evm::{EthEvmFactory, EvmFactory};
     use alloy_network::{AnyTxType, UnknownTxEnvelope, UnknownTypedTransaction};
-    use alloy_op_evm::OpEvmFactory;
     use alloy_primitives::Signature;
     use alloy_rpc_types::{Transaction as RpcTransaction, TransactionInfo};
     use alloy_serde::WithOtherFields;
     use foundry_evm_hardforks::TempoHardfork;
-    use op_alloy_consensus::{OpTxEnvelope, transaction::OpTransactionInfo};
-    use op_alloy_rpc_types::Transaction as OpRpcTransaction;
-    use op_revm::OpSpecId;
     use revm::database::EmptyDB;
     use tempo_alloy::primitives::{
         AASigned, TempoSignature, TempoTransaction, TempoTxEnvelope,
@@ -785,30 +792,6 @@ mod tests {
         // Test EVM Context Cfg mutation
         evm.ctx_mut().cfg_mut().spec = SpecId::AMSTERDAM;
         assert_eq!(evm.ctx().cfg().spec, SpecId::AMSTERDAM);
-
-        // Round-trip test to ensure no issues with cloning and setting tx_env and evm_env
-        let tx_env = evm.ctx().tx_clone();
-        evm.ctx_mut().set_tx(tx_env);
-        let evm_env = evm.ctx().evm_clone();
-        evm.ctx_mut().set_evm(evm_env);
-    }
-
-    #[test]
-    fn op_evm_foundry_context_ext_implementation() {
-        let mut evm =
-            OpEvmFactory::<OpTx>::default().create_evm(EmptyDB::default(), EvmEnv::default());
-
-        // Test EVM Context Block mutation
-        evm.ctx_mut().block_mut().set_number(U256::from(123));
-        assert_eq!(evm.ctx().block().number(), U256::from(123));
-
-        // Test EVM Context Tx mutation
-        evm.ctx_mut().tx_mut().set_nonce(99);
-        assert_eq!(evm.ctx().tx().nonce(), 99);
-
-        // Test EVM Context Cfg mutation
-        evm.ctx_mut().cfg_mut().spec = OpSpecId::JOVIAN;
-        assert_eq!(evm.ctx().cfg().spec, OpSpecId::JOVIAN);
 
         // Round-trip test to ensure no issues with cloning and setting tx_env and evm_env
         let tx_env = evm.ctx().tx_clone();
@@ -875,23 +858,6 @@ mod tests {
     }
 
     #[test]
-    fn from_any_rpc_transaction_for_op() {
-        let from = Address::random();
-        let signed_tx = make_signed_eip1559();
-
-        // Build the eth TxEnv to compare against op base
-        let rpc_tx = RpcTransaction::from_transaction(
-            Recovered::new_unchecked(signed_tx.into(), from),
-            TransactionInfo::default(),
-        );
-        let any_tx = <AnyRpcTransaction as From<RpcTransaction>>::from(rpc_tx);
-        let expected_base = TxEnv::from_any_rpc_transaction(&any_tx).unwrap();
-
-        let op_tx_env = OpTx::from_any_rpc_transaction(&any_tx).unwrap();
-        assert_eq!(op_tx_env.base, expected_base);
-    }
-
-    #[test]
     fn from_any_rpc_transaction_unknown_envelope_errors() {
         let unknown = AnyTxEnvelope::Unknown(UnknownTxEnvelope {
             hash: B256::ZERO,
@@ -913,39 +879,6 @@ mod tests {
 
         let result = TxEnv::from_any_rpc_transaction(&any_tx).unwrap_err();
         assert!(result.to_string().contains("unknown transaction type"));
-    }
-
-    #[test]
-    fn from_any_rpc_transaction_for_op_deposit() {
-        let from = Address::random();
-        let source_hash = B256::random();
-        let deposit = TxDeposit {
-            source_hash,
-            from,
-            to: TxKind::Call(Address::with_last_byte(0xCC)),
-            mint: 1111,
-            value: U256::from(200),
-            gas_limit: 21000,
-            is_system_transaction: true,
-            input: Default::default(),
-        };
-
-        // Build a concrete OpRpcTransaction, serialize to JSON, deserialize as AnyRpcTransaction.
-        let op_rpc_tx = OpRpcTransaction::from_transaction(
-            Recovered::new_unchecked(OpTxEnvelope::Deposit(Sealed::new(deposit)), from),
-            OpTransactionInfo::default(),
-        );
-        let json = serde_json::to_value(&op_rpc_tx).unwrap();
-        let any_tx: AnyRpcTransaction = serde_json::from_value(json).unwrap();
-
-        let op_tx_env = OpTx::from_any_rpc_transaction(&any_tx).unwrap();
-        assert_eq!(op_tx_env.base.caller, from);
-        assert_eq!(op_tx_env.base.kind, TxKind::Call(Address::with_last_byte(0xCC)));
-        assert_eq!(op_tx_env.base.value, U256::from(200));
-        assert_eq!(op_tx_env.base.gas_limit, 21000);
-        assert_eq!(op_tx_env.deposit.source_hash, source_hash);
-        assert_eq!(op_tx_env.deposit.mint, Some(1111));
-        assert!(op_tx_env.deposit.is_system_transaction);
     }
 
     #[test]
@@ -1003,5 +936,89 @@ mod tests {
         assert_eq!(tx_env.inner.gas_limit, 424242);
         assert_eq!(tx_env.inner.chain_id, Some(42431));
         assert_eq!(tx_env.fee_token, fee_token);
+    }
+
+    #[cfg(feature = "optimism")]
+    mod optimism {
+        use super::*;
+        use alloy_op_evm::{OpEvmFactory, OpTx};
+        use op_alloy_consensus::{OpTxEnvelope, TxDeposit, transaction::OpTransactionInfo};
+        use op_alloy_rpc_types::Transaction as OpRpcTransaction;
+        use op_revm::OpSpecId;
+
+        #[test]
+        fn op_evm_foundry_context_ext_implementation() {
+            let mut evm =
+                OpEvmFactory::<OpTx>::default().create_evm(EmptyDB::default(), EvmEnv::default());
+
+            // Test EVM Context Block mutation
+            evm.ctx_mut().block_mut().set_number(U256::from(123));
+            assert_eq!(evm.ctx().block().number(), U256::from(123));
+
+            // Test EVM Context Tx mutation
+            evm.ctx_mut().tx_mut().set_nonce(99);
+            assert_eq!(evm.ctx().tx().nonce(), 99);
+
+            // Test EVM Context Cfg mutation
+            evm.ctx_mut().cfg_mut().spec = OpSpecId::JOVIAN;
+            assert_eq!(evm.ctx().cfg().spec, OpSpecId::JOVIAN);
+
+            // Round-trip test to ensure no issues with cloning and setting tx_env and evm_env
+            let tx_env = evm.ctx().tx_clone();
+            evm.ctx_mut().set_tx(tx_env);
+            let evm_env = evm.ctx().evm_clone();
+            evm.ctx_mut().set_evm(evm_env);
+        }
+
+        #[test]
+        fn from_any_rpc_transaction_for_op() {
+            let from = Address::random();
+            let signed_tx = make_signed_eip1559();
+
+            // Build the eth TxEnv to compare against op base
+            let rpc_tx = RpcTransaction::from_transaction(
+                Recovered::new_unchecked(signed_tx.into(), from),
+                TransactionInfo::default(),
+            );
+            let any_tx = <AnyRpcTransaction as From<RpcTransaction>>::from(rpc_tx);
+            let expected_base = TxEnv::from_any_rpc_transaction(&any_tx).unwrap();
+
+            let op_tx_env = OpTx::from_any_rpc_transaction(&any_tx).unwrap();
+            assert_eq!(op_tx_env.base, expected_base);
+        }
+
+        #[test]
+        fn from_any_rpc_transaction_for_op_deposit() {
+            let from = Address::random();
+            let source_hash = B256::random();
+            let deposit = TxDeposit {
+                source_hash,
+                from,
+                to: TxKind::Call(Address::with_last_byte(0xCC)),
+                mint: 1111,
+                value: U256::from(200),
+                gas_limit: 21000,
+                is_system_transaction: true,
+                input: Default::default(),
+            };
+
+            // Build a concrete OpRpcTransaction, serialize to JSON, deserialize as
+            // AnyRpcTransaction.
+            let op_rpc_tx = OpRpcTransaction::from_transaction(
+                Recovered::new_unchecked(OpTxEnvelope::Deposit(Sealed::new(deposit)), from),
+                OpTransactionInfo::default(),
+            );
+            let json = serde_json::to_value(&op_rpc_tx).unwrap();
+            let any_tx: AnyRpcTransaction = serde_json::from_value(json).unwrap();
+
+            let op_tx_env = OpTx::from_any_rpc_transaction(&any_tx).unwrap();
+            assert_eq!(op_tx_env.base.caller, from);
+            assert_eq!(op_tx_env.base.kind, TxKind::Call(Address::with_last_byte(0xCC)));
+            assert_eq!(op_tx_env.base.value, U256::from(200));
+            assert_eq!(op_tx_env.base.gas_limit, 21000);
+            assert_eq!(op_tx_env.deposit.source_hash, source_hash);
+            assert_eq!(op_tx_env.deposit.mint, Some(1111));
+            assert!(op_tx_env.deposit.is_system_transaction);
+        }
     }
 }
