+++ crates/evm/core/src/evm/mod.rs
@@ -10,14 +10,11 @@ use alloy_evm::{
     EthEvmFactory, Evm, EvmEnv, EvmFactory, FromRecoveredTx, precompiles::PrecompilesMap,
 };
 use alloy_network::{Ethereum, Network};
-use alloy_op_evm::OpEvmFactory;
 use alloy_primitives::{Address, Signature, U256};
 use alloy_rlp::Decodable;
 use foundry_common::{FoundryReceiptResponse, FoundryTransactionBuilder, fmt::UIfmt};
 use foundry_config::FromEvmVersion;
 use foundry_fork_db::{DatabaseError, ForkBlockEnv};
-use op_alloy_network::Optimism;
-use op_revm::OpHaltReason;
 use revm::{
     Database,
     context::{
@@ -36,10 +33,12 @@ use tempo_evm::evm::TempoEvmFactory;
 use tempo_revm::TempoHaltReason;
 
 pub mod eth;
+#[cfg(feature = "optimism")]
 pub mod op;
 pub mod tempo;
 
 pub use eth::*;
+#[cfg(feature = "optimism")]
 pub use op::*;
 pub use tempo::*;
 
@@ -73,13 +72,6 @@ pub struct TempoEvmNetwork;
 impl FoundryEvmNetwork for TempoEvmNetwork {
     type Network = TempoNetwork;
     type EvmFactory = TempoEvmFactory;
-}
-
-#[derive(Clone, Copy, Debug, Default)]
-pub struct OpEvmNetwork;
-impl FoundryEvmNetwork for OpEvmNetwork {
-    type Network = Optimism;
-    type EvmFactory = OpEvmFactory;
 }
 
 /// Convenience type aliases for accessing associated types through [`FoundryEvmNetwork`].
@@ -246,15 +238,6 @@ pub trait IntoInstructionResult {
 impl IntoInstructionResult for HaltReason {
     fn into_instruction_result(self) -> InstructionResult {
         self.into()
-    }
-}
-
-impl IntoInstructionResult for OpHaltReason {
-    fn into_instruction_result(self) -> InstructionResult {
-        match self {
-            Self::Base(eth) => eth.into(),
-            Self::FailedDeposit => InstructionResult::Stop,
-        }
     }
 }
