+++ crates/evm/evm/Cargo.toml
@@ -18,7 +18,7 @@ foundry-cheatcodes.workspace = true
 foundry-common.workspace = true
 foundry-compilers.workspace = true
 foundry-config.workspace = true
-foundry-evm-core.workspace = true
+foundry-evm-core = { workspace = true, features = ["optimism"] }
 foundry-evm-coverage.workspace = true
 foundry-evm-fuzz.workspace = true
 foundry-evm-sancov.workspace = true
